(refactor) refactored jwtCheck to be in middleware
